backend/rss_importer/requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-slugify==8.0.4
psycopg2-binary==2.9.10
APScheduler==3.10.4
pytz==2024.2